Rename files; first example DB transactions
